api/requirements.txt
flask~=2.2.2
flask_restful
flask_cors
peewee~=3.15.4
flask_peewee
psycopg2-binary
peewee-migrate
werkzeug~=2.2.2
pika~=1.3.1
openai==0.26.0
python-magic~=0.4.27
flask-socketio~=5.3.2
textract~=1.6.3